requirements.txt
pytest
pytest-django
pytest-cov
psycopg2-binary
coverage[toml]
